+++ changelog.md
@@ -1,6 +1,7 @@
 - added Gambling Card
 - added Reverse Card
 - added new advancement for 69% efficiency bonus
+- added 25% chance for Swimmer Guide to spawn in buried treasures (suggested by @Ladocterus)
 - removed Jadeite
 - reworked Wandering Trader trades for accessories and boosters
 - updated Chinese translation (thanks to @Fiocher)